resolving bugs, removing uneccessary comments and console logs

+++ src/components/owner-dashboard.js
@@ -6,7 +6,7 @@ import { job_details } from './fetch';
 import PetsCard from './partials/pets-card';
 // import JobsCard from './partials/jobs-card';
 
-import { getUserData, userSession } from '../auth';
+import { userSession } from '../auth';
 
 let stxAddress;
 let str_stxAddress;
@@ -74,31 +74,32 @@ function JobsCard() {
         } else {
             visit = "Walk"
         }
-
-    jobsCardArray[i] = 
-        <div>
-            <div class="jobs">
-                <div class="card mb-2">
-                    <div class="card-body">
-                        <h4 class="card-title">{job_details().allJobs[i].j_owner_first_name} {job_details().allJobs[i].j_owner_last_name}</h4>
-                        <h5 class="card-text">Pet: {job_details().allJobs[i].j_pet_name}</h5>
-                        <h5 class="card-subtitle">Visit: {visit}</h5>
-                        <h5 class="card-text">Pay: {job_details().allJobs[i].j_pay}</h5>
-                        <h6 class="card-text text-muted">Time: {job_details().allJobs[i].j_timeframe}</h6>
-                        <h6 class="card-text text-muted">Location: {job_details().allJobs[i].j_location}</h6>
-                        {/* {{!-- (HELPER) function that determines if the person logged in is the creator of this job and it's active--}} */}
-                        {/* {{#unless completed}} */}
-                        <button id="delete-job" class="float-right btn success" type="button" onClick={() => DeleteJob(job_details().allJobs[i].j_id)}>Delete</button>
-                        {/* {{/unless}} */}
-                        {/* {{!-- (HELPER) if logged in as walker and job not accepted --}}
-                        {{#if acceptable}} */}
-                        {/* <button id="accept-job" class="float-right btn success" type="button">Accept</button> */}
-                        {/* {{/if}} */}
+    if (job_details().allJobs[i].j_owner_id === stxAddress) {
+        jobsCardArray[i] = 
+            <div>
+                <div class="jobs">
+                    <div class="card mb-2">
+                        <div class="card-body">
+                            <h4 class="card-title">{job_details().allJobs[i].j_owner_first_name} {job_details().allJobs[i].j_owner_last_name}</h4>
+                            <h5 class="card-text">Pet: {job_details().allJobs[i].j_pet_name}</h5>
+                            <h5 class="card-subtitle">Visit: {visit}</h5>
+                            <h5 class="card-text">Pay: {job_details().allJobs[i].j_pay}</h5>
+                            <h6 class="card-text text-muted">Time: {job_details().allJobs[i].j_timeframe}</h6>
+                            <h6 class="card-text text-muted">Location: {job_details().allJobs[i].j_location}</h6>
+                            {/* {{!-- (HELPER) function that determines if the person logged in is the creator of this job and it's active--}} */}
+                            {/* {{#unless completed}} */}
+                            <button id="delete-job" class="float-right btn success" type="button" onClick={() => DeleteJob(job_details().allJobs[i].j_id)}>Delete</button>
+                            {/* {{/unless}} */}
+                            {/* {{!-- (HELPER) if logged in as walker and job not accepted --}}
+                            {{#if acceptable}} */}
+                            {/* <button id="accept-job" class="float-right btn success" type="button">Accept</button> */}
+                            {/* {{/if}} */}
+                        </div>
                     </div>
+                    {/* {{/each}} */}
                 </div>
-                {/* {{/each}} */}
-            </div>
-        </div>;
+            </div>;
+        }
     }
     return jobsCardArray;
 }
@@ -138,7 +139,7 @@ function OwnerDash() {
                                 <button className="float-right mr-3 mb-3 btn success" type="submit" onClick={() => theBtn()}>Confirm</button>
                             </div>
                         </form>
-                    </div>  
+                    </div> 
                 </div>
                 
                 <div id="jobs-container" className="d-flex flex-column mx-auto" style={{width: '92%'}}>
